Reorganize & annotate observer.ts.
