+++ src/org/datafoodconsortium/connector/codegen/typescript/static/src/IConnectorExporter.ts
@@ -1,6 +1,6 @@
 import { Semanticable } from "@virtual-assembly/semantizer";
 import IConnectorExporterOptions from "./IConnectorExporterOptions";
-import { Observable } from "./Observable";
+import { Observable } from "./observer";
 
 export default interface IConnectorExporter extends Observable<string> {
